new member vcfbed; move variant_annotation.py and bed_writer.py
