+++ .github/utils/ugbio_members_matrix.py
@@ -3,7 +3,7 @@ import argparse
 import json
 import os
 
-INTERNAL_MEMBERS = ["core"]
+INTERNAL_MEMBERS = ["core", "vcfbed"]
 
 
 def build_members_matrix(src_folder: str, *, keep_internal: bool):
