+++ src/cnv/pyproject.toml
@@ -10,7 +10,6 @@ description = "Ultima Genomics CNV utils"
 authors = [
     { name = "Tammy Biniashvili", email = "tammy.biniashvili@ultimagen.com" },
     { name = "Avigail Moldovan", email = "avigail.moldovan@ultimagen.com" },
-    { name = "Tammy Biniashvili", email = "tammy.biniashvili@ultimagen.com" },
 ]
 readme = "README.cnv.md"
