+++ src/ppmseq/pyproject.toml
@@ -3,7 +3,8 @@ name = "ugbio_ppmseq"
 version = "1.3.5-0"
 requires-python = ">=3.11"
 dependencies = [
-    "ugbio_core[variantannotation]",
+    "ugbio_core",
+    "ugbio_vcfbed",
     "seaborn>=0.13.2",
     "papermill>=2.6.0",
     "jupyter>=1.1.1",
@@ -34,6 +35,9 @@ dev-dependencies = [
 ]
 
 [tool.uv.sources.ugbio_core]
+workspace = true
+
+[tool.uv.sources.ugbio_vcfbed]
 workspace = true
 
 [tool.setuptools.package-data]
