+++ src/srsnv/pyproject.toml
@@ -3,7 +3,8 @@ name = "ugbio_srsnv"
 version = "1.3.5-0"
 requires-python = ">=3.11"
 dependencies = [
-    "ugbio_core[variantannotation]",
+    "ugbio_core",
+    "ugbio_vcfbed",
     "ugbio_ppmseq",
     "ugbio_featuremap",
     "seaborn>=0.13.2",
@@ -42,6 +43,9 @@ dev-dependencies = [
 ]
 
 [tool.uv.sources.ugbio_core]
+workspace = true
+
+[tool.uv.sources.ugbio_vcfbed]
 workspace = true
 
 [tool.uv.sources.ugbio_ppmseq]
