+++ src/vcfbed/README.vcfbed.md
@@ -1,0 +1,3 @@
+# ugbio_vcfbed
+
+This module includes VcfBED python scripts for bioinformatics pipelines.
