+++ src/vcfbed/pyproject.toml
@@ -1,0 +1,36 @@
+[project]
+name = "ugbio_vcfbed"
+version = "1.3.5-0"
+requires-python = ">= 3.11"
+dependencies = [
+    "ugbio_core",
+    "pybigwig>=0.3.18",
+    "tqdm>=4.66.4",
+    "bgzip>=0.5.0"
+]
+description = "Ultima Genomics VcfBED utils"
+authors = [
+    { name = "Avigail Moldovan", email = "avigail.moldovan@ultimagen.com" },
+    { name = "Tammy Biniashvili", email = "tammy.biniashvili@ultimagen.com" }
+]
+readme = "README.vcfbed.md"
+
+[project.scripts]
+annotate_contig = "ugbio_vcfbed.annotate_contig:main"
+run_tests = "pytest:main"
+
+[tool.uv]
+dev-dependencies = [
+    "pre-commit>=4.0.1",
+    "pytest>=8.3.3",
+    "ruff>=0.6.9",
+]
+
+[tool.uv.sources.ugbio_core]
+workspace = true
+
+[build-system]
+requires = [
+    "setuptools>=61.0",
+]
+build-backend = "setuptools.build_meta"
