+++ src/vcfbed/tests/resources/sample.fasta
